Modified most ids to be represented as int instead of long so that persistence to data files takes up less space.  This will only work until the database reaches 2 billion'ish nodes ...

+++ src/com/bretth/osmosis/core/bdb/v0_5/impl/DatabaseTileIndexIterator.java
@@ -86,7 +86,9 @@ public class DatabaseTileIndexIterator implements ReleasableIterator<Long> {
 			
 			// We want to start retrieving records from the first record with a
 			// part 1 of tileId. Therefore part2 must be the minimum value.
-			startKey = new UnsignedIntegerLongIndexElement(minimumTile, Long.MIN_VALUE);
+			// HACK: Note that this should be using Long.MIN_VALUE but the
+			// internal implementation is currently using an integer.
+			startKey = new UnsignedIntegerLongIndexElement(minimumTile, Integer.MIN_VALUE);
 			keyBinding.objectToEntry(startKey, keyEntry);
 			
 			cursor = db.openCursor(txn, null);
